refactor `BossStationPatch`
